fix on pipe low level semantics

+++ include/pipe.h
@@ -8,7 +8,7 @@ struct pipe_s *pipe_create(uint16_t size);
 int32_t pipe_destroy(struct pipe_s *pipe);
 void pipe_flush(struct pipe_s *pipe);
 int32_t pipe_size(struct pipe_s *pipe);
-int8_t pipe_get(struct pipe_s *pipe);
+int32_t pipe_get(struct pipe_s *pipe);
 int32_t pipe_put(struct pipe_s *pipe, int8_t data);
 int32_t pipe_read(struct pipe_s *pipe, int8_t *data, uint16_t size);
 int32_t pipe_write(struct pipe_s *pipe, int8_t *data, uint16_t size);
